+++ tests/test.py
@@ -1,5 +1,4 @@
 import os, traceback
-import signal 
 from pymongo import MongoClient
 from pymongoose.methods import set_schemas, get_cursor_length
 from pymongoose.methods import AS_DICT, AS_STRING
@@ -139,6 +138,32 @@ def test_find():
         traceback.print_exc()
         raise Exception("Error initializing database")
         return 1
+
+# def test_find_no_timeout ():
+#     mongo_init()
+#     try:
+#         users = User.find({}, no_cursor_timeout=True)
+#         users_len = 0
+
+#         users: Cursor
+
+#         for user in users:
+#             users_len += 1
+
+#         users.close ()
+
+#         print ("Closed??")
+        
+#         if users is None or users_len == 0:
+#             return 1
+
+#         print("Find exit with code 0")
+#         return 0
+
+#     except:
+#         traceback.print_exc()
+#         raise Exception("Error initializing database")
+#         return 1
 
 def test_find_dict():
     mongo_init()
@@ -471,6 +496,7 @@ def test_delete_many():
 
 # test_insert()
 # test_find()
+# test_find_no_timeout ()
 # test_find_dict()
 # test_find_string()
 # test_find_sort()