+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ## Methods
-- Fixed how all parents are grouped when joining children in populate
+- Added no_cursor_timeout option in find methods